LayoutTests/editing/selection/ios/tap-focused-input-clears-outside-selection.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1">
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/js-test.js"></script>
<style>
.target {
    font-size: 40px;
    width: 300px;
}

input {
    display: block;
    margin-top: 40px;
    width: 300px;
    height: 80px;
    font-size: 40px;
}
</style>
<script>
jsTestIsAsync = true;
description("Verifies that tapping a previously-focused input while a non-editable selection is active refocuses the input and clears the outside selection (mirrors UIKit TextField behavior: tapping the field reclaims first-responder and discards any standalone selection). Requires WebKitTestRunner.");

addEventListener("load", async () => {
    const input = document.querySelector("input");

    // Set up: focus the input (keyboard up), then long-press outside to pull focus away
    // and create a non-editable selection. After this, the input is blurred and 'Hello' is selected.
    const inputRect = input.getBoundingClientRect();
    await UIHelper.activateAndWaitForInputSessionAt(inputRect.left + inputRect.width / 2, inputRect.top + inputRect.height / 2);
    await UIHelper.longPressElement(document.querySelector(".target"));
    await shouldBecomeEqual("getSelection().toString()", "'Hello'");

    // Now tap the (currently blurred) input. UIKit's behavior: the field becomes first responder again,
    // and the standalone selection is discarded.
    await UIHelper.activateElementAndWaitForInputSession(input);

    window.focusedElementIsInput = document.activeElement === input;
    shouldBeTrue("focusedElementIsInput");

    // The standalone 'Hello' selection must have been discarded; the document-level selection
    // should no longer reflect any outside text.
    window.outsideSelectionGone = getSelection().toString() === "";
    shouldBeTrue("outsideSelectionGone");

    // And the caret should now live inside the input as a collapsed selection on its internal
    // text control. (The input owns its selection internally, not via document.getSelection().)
    window.inputCaretCollapsed = input.selectionStart === input.selectionEnd;
    shouldBeTrue("inputCaretCollapsed");

    finishJSTest();
});
</script>
</head>
<body>
    <span class="target">Hello</span>
    <input>
</body>
</html>
