LayoutTests/editing/selection/ios/tap-focused-input-clears-outside-selection-expected.txt

Verifies that tapping a previously-focused input while a non-editable selection is active refocuses the input and clears the outside selection (mirrors UIKit TextField behavior: tapping the field reclaims first-responder and discards any standalone selection). Requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getSelection().toString() became 'Hello'
PASS focusedElementIsInput is true
PASS outsideSelectionGone is true
PASS inputCaretCollapsed is true
PASS successfullyParsed is true

TEST COMPLETE
Hello
